[!!!][FEATURE] Automatically fetch scope domain (#7)
Uses other domains we have in the configuration to determine it. This means the scope string doesn't have to be specified by the integrator when configuring the extension.

The scope domain is where the CNAME record of the oceDomain points. If no CNAME record exists, the oceDomain is used, as it logically follows that a domain without CNAME record cannot be resolved further and thus must be the scope domain.